backend/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
python-dotenv
google-generativeai
requests
gdown